fix: adjust function position
